Provider clients join core/apis; ticketing keeps only tracker policy
